tui/requirements.txt
pyserial>=3.5
npyscreen>=4.10.5
windows-curses; platform_system == "Windows"
